+++ src/App.tsx
@@ -5,12 +5,12 @@ import { ListingPage } from 'pages/ListingPage'
 import { SummaryPage } from 'pages/SummaryPage'
 import { LoginPage } from 'pages/LoginPage'
 import { Toast } from 'components/Toast'
+import { AuthProvider } from 'contexts/authContext'
 
 import { ThemeSwitcherProvider } from './hooks/useThemeSwitcher'
 import { TransactionsProvider } from './hooks/useTransactions'
 
 import { GlobalStyle } from './styles/global'
-import { AuthProvider } from 'contexts/authContext'
 
 Modal.setAppElement('#root')
 
@@ -18,17 +18,17 @@ export function App() {
   return (
     <BrowserRouter>
       <AuthProvider>
-          <ThemeSwitcherProvider>
-            <Toast />
-            <TransactionsProvider>
-              <Switch>
-                <Route exact path="/" component={ListingPage} />
-                <Route path="/summary" component={SummaryPage} />
-                <Route path="/auth" component={LoginPage} />
-              </Switch>
-              <GlobalStyle />
-            </TransactionsProvider>
-          </ThemeSwitcherProvider>
+        <ThemeSwitcherProvider>
+          <Toast />
+          <TransactionsProvider>
+            <Switch>
+              <Route exact path="/" component={ListingPage} />
+              <Route path="/summary" component={SummaryPage} />
+              <Route path="/auth" component={LoginPage} />
+            </Switch>
+            <GlobalStyle />
+          </TransactionsProvider>
+        </ThemeSwitcherProvider>
       </AuthProvider>
     </BrowserRouter>
   )
