+++ src/domain/NavigationBar/styles.ts
