+++ src/domain/RemoveTransactionDialog/styles.ts
@@ -2,13 +2,13 @@ import styled from 'styled-components'
 
 export const Container = styled.form`
   h2 {
-    color: ${ctx => ctx.theme.colors.textTitle};
+    color: ${(ctx) => ctx.theme.colors.textTitle};
     font-size: 1.5rem;
     margin-bottom: 2rem;
   }
 
   p {
-    color: ${ctx => ctx.theme.colors.text}
+    color: ${(ctx) => ctx.theme.colors.text};
   }
 
   .actions {
@@ -18,7 +18,6 @@ export const Container = styled.form`
 
     button {
       flex: 1;
-      padding: 0 1.5rem;
     }
   }
 `
