+++ src/domain/TransactionsTable/TransactionCard/index.tsx
