+++ src/domain/TransactionsTable/TransactionCard/styles.ts
